reddit-highrisk-scraper/requirements.txt
praw
pandas
pyyaml
requests
python-dotenv
